+++ src/ModFramework/obj_database.lua
@@ -3,24 +3,23 @@
 ---@param q game_obj_database
 ---@param v_modid string
 function create(q,v_modid)
-	--path to the current file
-	local currentFilePath = (mod_info[v_modid]):gsub("obj_database.lua","")
-
-	--Add current mod path to the package.path
-	--This ensures we can load the mod framework in every file in every mod
-	local modFramework = ";"..currentFilePath.."ModFramework\\?.lua"
-	local modFrameworkInternal = ";"..currentFilePath.."ModFrameworkInternal\\?.lua"
-	package.path = package.path..modFramework..modFrameworkInternal
-
 	--Only needed in the framework setup, is not needed for mods
 	Internal = require("ModFrameworkInternal")
+
 	Internal.RegisterFramework()
+	Internal.Settings.LoadMenuSprites()
+	Internal.Engineering.LoadMenuSprites()
 	Internal.ComponentShop.LoadShopSprites()
+
+	Internal.Settings.RegisterBooleanSetting("QuickMovePilots", true, { LocalizedDefaultValue = "Allow pilots to be moved by using shift + left mouse button" })
+	Internal.Settings.RegisterBooleanSetting("RTSSelectionStyle", false, { LocalizedDefaultValue = "Replace the battle selection methods with an RTS style one" })
+	Internal.Settings.RegisterBooleanSetting("PartialEngineering", false, { LocalizedDefaultValue = "Allow engineers to do partial mech assembly" })
 end
 
 ---saving system deletes the file and creates new one before saving new info
 ---@param q game_obj_database
 function save_game_pre_event(q)
+	Internal.Settings.SaveData()
 end
 
 ---@param q game_obj_database
@@ -34,6 +33,7 @@ end
 ---Called after the game is loaded
 ---@param q game_obj_database
 function load_game_post_event(q)
+	Internal.Settings.LoadData()
 	Internal.Production.SetModdedSprites()
 end
 
@@ -46,6 +46,7 @@ function draw_top_menu(q)
 	Internal.ComponentShop.FixRobotComponentBleed()
 	Internal.Research.FixModdedResearch()
 	Internal.Hanger.ProcessPilotDataQueue()
+	Internal.Settings.DrawMenu()
 end
 
 ---The draw call that runs every frame while debug is active (F6)